God-mode drop of loot into building.
Extend the god-mode "drop loot" command to allow dropping loot
also inside a building (into the inventory of a given account there).
This is useful for testing, especially since limited cargo space of
vehicles means that it may require a lot of action to "transport" loot
dropped onto the ground into a building.

The new command looks like this:

  {"drop": [{"building": {"id": 123, "a": "domob"}, "fungible": {"foo": 42}}]}