Implement "repair armour" service operation.
The new "repair armour" service operation requests that the armour of the
vehicle of a character is repaired, costing some vCHI and taking some
blocks to do (both depending on the missing number of HPs).

The new service operation can be requested with a move like this:

  {"s": [{"t": "fix", "b": 10, "c": 20}]}

The character (20 in this case) must be owned by the account making
the request, and must be inside the building where the service is
requested (10 in the example).