Support services fees in buildings.
If configured in a building, pay an additional service fee to the
building owner when someone else uses services.  For neutral buildings
and your own buildings, there is never such a fee.

For now, the fee is handled by the service logic if configured, but there
is not yet a way to actually configure the fee (except in tests).  This will
come as a next step.