Fix invalidated iterator bug.
One test was using invalidated iterators, because it was iterating
over an inventory while removing the items at the same time.  This was
causing memory issues.

This change introduces a Clear method for inventories, so that we can just
clear the whole inventory afterwards directly.  With this, all tests
run fine in valgrind without errors.