Test for order between services and char updates.
This adds a new unit test for the relative order of processing
service requests and character updates in a move (services are after
character updates).

With the repair service, the order may now matter (e.g. if a character
is requesting armour repair and leaving the building it is in at the
same time).