Introduce RoItemData method.
Instead of looking up roconfig item data directly in the proto's
fungible_items map, a new function RoItemData() should be used to do that.

For now, the function just does the map lookup; but in the future it will
also handle "programmatic item types" that are not present per se in the
roconfig data, like blueprints or tech levels.