Generalise prize count table.
Instead of having a database table specifically for number of found
prizes, change this to a table that can count found items of general
type (and doesn't need to be initialised to all zeros).

We will need that for blueprints as well, to modify the chances of finding
more of one type based on how many there are already.