Implement reverse-engineering service.
The new service type reverse engineers one or more artefacts in a building,
with a chance to get a blueprint instead.  The cost and the possible outcomes
is configured per artefact type.  The chance for finding a blueprint
decreases with the number of existing blueprints of that type in the game
(with a factor of 75% for each existing blueprint).

The service can be requested with a move like this (similar to refining):

  {"s": [{"t": "rve", "b": 123, "i": "artefact", "n": 1}]}