Use ongoing operations for character busy.
This replaces the "old" busy-concept for characters (used for prospecting
and armour repair) with the new "ongoing operations".  Characters will
now only store the ongoing-ID in their proto to indicate that they are
currently busy.  Processing of finished operations is now done through
the general system instead of the previous DecrementBusy.