Building service for blueprint copy.
Define the building service for copying blueprints.  This takes one original
blueprint and produces one or more copies, using up a certain amount of
vCHI and taking a certain amount of blocks.  The cost and duration are
proportional to the blueprint's base item's "complexity" measure.

To request this service, the move format is:

  {"s": [{"b": 42, "t": "cp", "i": "item bpo", "n": 10}]}