Define construction service operation.
This defines a new service operation, which can be used to construct
either items or vehicles from blueprints.  (These two types of construction
are different building services, construction facility vs vehicle bay.)

Items can be constructed from a blueprint original or blueprint copy.
In either case, they cost a certain amount of resources (defined by the
base item type) per constructed item, and a certain amount of vCHI defined
by the base item's complexity.  Blueprint copies are used up, originals
will just be temporarily taken away (and then given back again).

When items are constructed from an original, they are done "in series",
i.e. it takes Nx the base number of blocks to construct N items (and the
base duration depends on the item complexity).  When constructed from
blueprint copies, the user needs to have N copies available anyway, but then
construction is done "in parallel" so it only takes the base number of
blocks to do it.

The new service can be requested with this move:

  {"s": [{"b": 42, "t": "bld", "i": "item bpo", "n": 5}]}

The "i" field is passed the blueprint that should be used (not the base item).
Based on that, the GSP figures out whether or not this is a construction from
original, and whether it is a vehicle or fitment (normal item).