Define starter vehicles as items.
Instead of defining and initialising the starting vehicle stats
in a custom Params function, define them as actual item types in the
roconfig data.

We also created a function to "derive" the character stats from the
character's vehicle item, which we use to initialise them now.  At the
moment, this just copies some proto fields.  In the future, this function
will also take fitments into account to modify derived stats accordingly.