Move for changing fitments.
There is a new move for setting the fitments of a character/vehicle.
This is only possible while inside a building and when shield and armour
are fully healed; also all fitments placed must either be already on
the vehicle or in the building inventory.  When placing fitments, all
of the character inventory is auto-dropped into the building (to avoid
issues with removed cargo expanders or things like that).

The move to set fitments looks like this:

  {"c": {"1": {"fit": ["expander", "multiplier", "cannon", "cannon"]}}}

It simply changes the fitments to the precise list as given, replacing
any fitments currently on the vehicle (either reusing them if listed
again or dropping into the building inventory if not).

If some of the fitments cannot be placed (e.g. because of hitting the
complexity limit, missing the item in the inventory or lack of slots),
then the entire move is ignored.