Restructure ItemData proto with oneof.
The ItemData proto has a couple of fields specific to item types
(e.g. stats for vehicle items or data about refining the material).

These are conceptually already a oneof, but were not in the actual code.
This change makes them a oneof.  This structures the proto more clearly,
and also gives an easier-to-see separation between some fields that are
general for items and some that define specific types of items.