Move for changing of vehicles.
A new move allows changing of the vehicle for a character.  This works
inside a building and if the character has full HP.  The new vehicle
must be in the account's building inventory.

When the change is successful, all inventory of the character will be
dropped inside the building, and all fitments will be undone.  (It is
however possible to re-equip fitments and pickup inventory after
changing the vehicle in a single move.)

The move is this:

  {"c": {"1": {"v": "vehicle type"}}}