Foundation concept in protos.
Extend the building protos (configuration and actual game-state data)
to include a concept of building foundations.  Buildings on the map
can now be only foundations, which will be reflected in the game state
and returned in JSON.

This also defines building roconfig data for how HPs look like in the
foundation vs full building, and what the stats (resources and time
required) are for building some type of building.