Implement CanPlaceBuilding logic.
This adds a new function, CanPlaceBuilding, which implements the checks
for whether or not a building can be placed in a certain configuration.

This is possible unless there is any kind of obstacle (base map or
dynamic) on any building tile, and only if the entire building is contained
within one region of the map.