Move processing for laying foundations.
This implements move processing for building the foundation of a new
building.  For that, the required resources must be in a character's
inventory and the character must not be busy.

When successful, the foundation is created with centre on the character's
position, and the character ends up inside.

The move format is this:

  {"c": {"1": {"fb": {"t": "building type", "rot": 3}}}}