Initialise DynObstacles for move processor tests.
This initialises a fresh DynObstacles instance from the actual database
state before running move processor tests.  Previously we just used a dummy
instance initialised at the start of the test, which meant that nothing
was on it.

In order to allow the proper initialisation to pass, we have to make sure
that test characters and buildings do not actually overlap in a way that
is not valid.