Add DynObstacles to pending move tracker.
For validating "found building" moves, we need an instance of the
DynObstacles object.  Since this is costly to construct (unlike the
database-table wrappers), we add an instance that is shared for processing
multiple moves in the pending tracker, and only recreated when the
confirmed state changes (because then its state will change as well).