Allow Inventory to wrap existing proto.
Extend the Inventory class (which is used in game code to interact with
inventories) so that it can not only hold a database value thorugh
LazyProto, but that it can also be used to just wrap an existing proto
by reference.  This is useful e.g. to work with inventories that are
just a field of another proto and not their own database column.