Add getbuildingshape RPC.
The new getbuildingshape RPC is stateless (i.e. works without locking
the game state, and is local with Charon).  It simply returns the
shape tiles for a given building type, centre and rotation.

This will be used by the frontend while placing a building that is not
yet on the map.

Fixes https://github.com/xaya/taurion_gsp/issues/104.