Support zero attack range.
Previously, an attack range value of zero (both in the code and in the
database) would mean that an entity has no attacks.  (In contrast to comments
in the SQL schema, that said that NULL means that.)

This adjusts the code to actually use NULL in the database and a named
constant "magic value" (of -1) in the code for this, so that we can also
properly support zero-range attacks (i.e. enemies on the same tile).

Maybe we won't need that, but the new, more explicit handling of "no attacks"
is cleaner anyway.