Allow AoE around target.
This restructures the way attack ranges and AoE ranges are defined
in the proto configuration.  Instead of having a single range and
a boolean "is AoE" field, we now have two ranges.  This allows for
three main types of attacks:

range set to N, area unset:  Simple attack with range N.
range unset, area set to N:  AoE around attacker with range N.
range set to R, area set to A:  Target within range up to R, and then do AoE
  around that target up to A tiles away.

The third option was not possible before, and is now implemented properly
in the core code and also used for some fitment items.