Simplify combat processing.
This refactors the logic processing combat damage.  Instead of processing
first all non-area and then all area attacks, and instead of summing up
the non-area damage before applying it, we simply go through each attack
in order now and process it.

That's the most straight-forward way to do it, and will also make it
simpler to adjust for applying effects (like slowing) instead of damage.

Note that this changes the order of random rolls and thus forks the state,
but does not otherwise change the "functionality".