Add LazyProto::IsEmpty method.
The new method LazyProto::IsEmpty returns whether or not it is known
that a message is empty, because it has been SetToDefault and then
not modified at all (or been initialised from an empty string and
then not been modified).

This is useful if we want to set a database column to NULL in the
case of an empty value, instead of storing an empty string.  We will
use that with combat effects.