Database storage for combat effects.
This introduces a new CombatEffects proto, which will store temporary
effects onto a character from combat (e.g. slowing or increased
range due to friendlies).

The character table has a new column to store those effects, with a quick
way to clear all effects (at the end of each turn).

The effect data is not yet filled in anywhere or used.