Apply effects during combat.
When processing combat attacks, also apply effects in addition to applying
damage (and reset the effects at the end of processing a block).

In addition to the general framework, this also defines two new fitment
attacks:  A retarder, which slows enemies in AoE, and a longretard, which
slows enemies in AoE around a selected target.