Update for Charon Server changes.
This updates the Charon server integrated with tauriond for the
recent changes in https://github.com/xaya/charon/pull/11.  To take
full advantage of the change, we also make sure that the Charon server
is running its own "main loop" which tries to reconnect periodically
if the XMPP connection gets closed.