Define low-HP boosts in combat data and fitments.
This adds a new concept of "low-HP boosts" to the fitments and also the
combat data of a character that has such a fitment equiped.  We also
define a fitment item already for it.

For now, it does not have any actual effect; in the future, these boosts
will increase damage and range of a character when it is low on armour HP.