Split validation from ServiceOperation::Parse.
This removes the validation from ServiceOperation::Parse, and
instead adds a new IsFullyValid method.  This has no functional
impact for now, but will allow us to retrieve the JSON representation
of a service later on with a new RPC even if the account balance is not
enough.  This will be used to retrieve associated fees in the
frontend.

squash: split out more validation