GetCustomStateData variant with raw Database.
This adds a new GetCustomStateData variant to PXLogic, which exposes
the raw Database instance.  It will be useful for RPC methods that do
not go through the GameStateJson abstraction, e.g. for extracting
information about a possible service operation.  (Which is state based,
but not part of the actual game state.)