Turn RoConfigData into a wrapper class.
Instead of a dumb function RoConfigData, we now have a wrapper class
RoConfig which exposes the underlying proto data.  This will allow
us in the future to associate the data to a chain (and have different
data for regtest), and it will also be a nice venue to add in specific
access functions for buildings and items.