Make RoConfig depend on Chain.
Make the RoConfig constructor take a xaya::Chain value, and update all
code to properly pass it through.

For now, this has no effect, but we will use it later to make the roconfig
depend on the chain (e.g. add test items and buildings only for regtest,
and move some/all of the Params values there).