Use REGTEST by default in ContextForTesting.
Switch the ContextForTesting to the REGTEST chain by default (and only
set it to MAINNET selectively in some tests as needed).  If we make the
roconfig depend on the chain, the test items used widely will only
be available on REGTEST, and thus most tests will need to be run
on the REGTEST configuration.