Move dev address to roconfig.
Migrate the developer address for character payments from Params to
roconfig.  This allows us to get rid of the duplicated definition in
the Python tests and instead use it directly from the protocol buffer.

Since the address differs on testnet from mainnet, we have to add
new support for a testnet-specific merge proto.