Use set for tracking kills.
When dealing damage and processing kills, use a set (with the TargetKey
helper class instead of the TargetId proto as value) instead of
a vector.

This allows us to more explicitly enforce that elements are unique,
and also provides a well-defined order by itself.  Both make the code
more robust and clearer.

Note that this does not have an actual in-game effect, because we already
did not add duplicated entries (it was just not cross-checked), and because
the order in which kills are processed does not matter (at least now).