Store prospecting blocks in character data.
Instead of defining the number of blocks prospecting needs in the general
game params, make it into a property stored for each character and
vehicle type.

With this, we can have vehicle types that do not support prospecting
at all, and we can also have different rates between different types
of vehicles.  Later on, we can even define fitments that alter the
prospecting rate.