Add DynObstacles instance to NonStateRpcServer.
Keep a DynObstacles instance in the NonStateRpcServer, and use it for
the edge weight in findpath.  For now, the instance is always empty
(so that this has no effect).  In the future, however, we will allow
to place buildings from the real game state into it through a second RPC.