Rework fitment configurations.
Adjust the fitment definitions in roconfig.  Now all of them are named
starting with "lf" (for "light fitment"), they are ordered as in the game
design sheet, and they have the real stats (except for resources and
space usage) of the game design.  Medium/heavy/very heavy fitments are
not yet defined, as well as certain types of effects that need to be
implemented in the game logic still.