Add armour/shield damage factors to config.
Allow the configuration of an attack to contain factors for how damage
affects armour and shield differently.  Two new fitments (laser beam
and rail gun) use this, which are specifically effective against
shield and armour, respectively.  (So far only the light variants
of the new fitments are included; we will update the data file completely
in a follow-up.)

The new fields are not yet taken into account when dealing damage,
which will be the next step.