Allow finding of artefacts when prospecting.
Actually check the chances and allow players to find artefacts when
they finish prospecting.  The logic is very similar to how prizes are
given out, namely just by checking against a list of possible outcomes
with chances.  (The outcomes and chances are based on what type of
resource is found at the position.)

Unlike prizes, artefacts do have a cargo size.  Thus it can happen that
they cannot be added to the character's inventory (because it is full),
in which case they will be dropped on the floor instead.