Remove wpdist from findpath.
Instead of spacing the waypoints by their distance (using the wpdist
argument), space them such that there is a principal direction between
each pair of waypoints.

This is what the GSP will need after the next change, which will simplify
the consensus-relevant movement code to not use path finding itself.