Movement along principal directions.
This revamps how the consensus-layer movement logic works in the GSP:
Instead of doing path finding in a local range, we expect the waypoints to
be set in principal directions from each other (but with arbitrary distance).
Then we simply try to step along that path.

This implements the idea described in
https://github.com/xaya/taurion_gsp/issues/135 and will make movement
processing (which has been the main bottleneck in the past two competitions)
much faster in the consensus code.  (Path-finding still needs to be done,
but only client side to come up with the waypoints to send in a move.)