moveTowards helper method in pxtest.
This adds a new helper method to the Character class in the integration
tests, which does a findpath call to obtain the path towards a given
target and then sets it as waypoints.

This decouples test logic that just needs to do some movement (without
caring about the real specifics) from how the waypoints need to be
specified (e.g. in principal directions).