Make blueprint copies and construction quick.
Set the cost and time required for blueprint copies and item construction
to minimal values (1 vCHI and 1 block per complexity) on mainnet, which
will help with the competition.

On regtest we now set it explicitly to the levels we had before, so that
the tests still work (and are testing a somewhat non-trivial situation
rather than all ones).