Update number of ore found on regions.
Instead of specifying a base amount of ore found on a prospected
region per type, we just specify a minimum and maximum value to be
found (for all types of ore) in the roconfig params.  Rare materials
will be rare through fewer regions and less output of refinement rather
than through less ores on any region.  (Although the random span of
outcomes is now much larger than before with the [X, 2X] system.)

This also updates the numbers to match the new units measured in cm^3,
namely to much higher values than before (2M to 100M for now and the
upcoming competition).