Define gain_hp function for fitments in roconfig.
This adds a new proto field for attacks, gain_hp.  When set, it means that
HP from damage of this attack will be added back to the attacker's own
HP set.  This will be used by the syphon and AoE syphon fitments (with
shield-only damage).

The change also defines the "lf" variants of the syphon fitments for testing
(the other variants will be added in later with a general update of the
fitment data).

For now the new field has no effect, but it will be implemented in the
combat logic in a follow-up.