Implement gain_hp in combat logic.
Properly handle HP gains (from attacks with gain_hp: true, i.e. the
syphon fitments) in the combat logic.  This is carefully done in a way
to not depend on the order in which characters are processed; gained HP
are credited only after processing all damage and won't save some
character from dying.

If more than one attacker gained from some target and the target runs dry
on HP to drain, then the HP are deducted but noone gets any credited; this
also helps to avoid any order dependence.

Most of the code is general and would support drains of armour as well,
but the syphon fitment does not need that (with armour damage percent set
to zero) and it would create some more ordering issues, so the code
asserts only shields are drained for now.