Fitment attribute for damage reduction.
This adds a new fitment attribute, which defines a modifier that is
applied to received damage.  This attribute gets passed through to a
character's CombatData, but there it is not yet taken into account when
actually processing the combat logic.

With this, we define the new "dmgred" fitment type (armour hardening)
in the "lf" variant for now.