Update protos for regeneratable armour.
Update the protocol buffer definitions and roconfig data to allow also
armour to regenerate (i.e. the accumulated mhp field is now a general HP
message, as is the regeneration_mhp field).

The code is not yet updated accordingly (and hence even breaks), nor is
there anything that actually *has* regenerating armour for now.