Implement range reduction combat effect.
This adds a new combat effect, which modifies (e.g. reduces) the range
of a targeted fighter.  This ability is given by a new fitment,
rangered, which reduces ranges with AoE style by 15%.

For now, we've defined the light variant of the new fitment.  Others will
be added later with a general data update.