Process effect updates after dealing damage.
Instead of clearing all effects at the end of block processing, we will
need to keep them active until after the next damaging phase so that e.g.
future range effects will take place.  (Conceptually, targeting, waiting
for the new block and then damaging is "one thing" anyway, so it makes
sense to not split it in the middle for how effects are handled.)

This change moves clearing and updating effects to right after damaging;
we accumulate the new effects in a temporary map (so that we keep the
old effects around as they are still what should be "active"), and then
swap them in after the damaging round inside the combat logic (but e.g.
before HP regeneration).