Add combat effects for buildings.
Actually store real combat effects for buildings in the database (with
the corresponding changes to the database schema and wrapper class),
and "move" the ClearAllEffects call to FighterTable (clearing both the
characters and buildings).

Previously, only speed changes (from a retarder) were implemented, which
obviously do not need to be there for buildings.  But future effects like
reduction in range will be applicable to buildings as well.