Define mobile refinery.
This extends the protocol buffers to support stats for a mobile refinery.
A fitment can have those, and then they will be propagated onto a
character that equips it (and returned in the character's JSON state).

So far, this does not do anything, but actually using it will be the
next step.

This also defines the "vhf" variant of the mobile refinery as actual
fitment, with the real stats.  That's the only one that will exist
(no lighter ones are possible).