Refining outputs should be smaller than inputs.
Add a unit test (and adjust the test ore) to ensure that the cargo
space used up by refining outputs is always less than the cargo
space used by the inputs.

This is something that should be true in general for game design, but
it is also crucial to ensure mobile refining works without issues
(where the inputs and outputs are based on a cargo-limited character
inventory rather than an unlimited building inventory).