Add allyreplenish fitment.
This defines a new fitment "lf allyreplenish", which increases the shield
regeneration rate of friendlies in an AoE.  For this, we define a new
combat effect for modifying the shield regeneration rate, and a flag for
attacks to mark them as "affects friendlies" (which is for simplicity
only supported with plain AoE).

Neither the shield regeneration modifier nor the actual friendly attacks
are implemented yet in the combat logic.