Support excluding buildings in findpath.
This adds a new argument "exbuildings" to the findpath RPC method.
It expects an array (might be empty) of building IDs (as previously
set with setpathbuildings).

Those buildings will then be ignored in the obstacle layer, i.e.
selectively not treated as obstacles.  This is useful for computing
paths "to" a building for entering it.