Store building IDs in setpathbuildings.
When setpathbuildings is called to define the buildings for findpath,
build up a map from tiles to the associated building ID in addition
to just the DynObstacle map.  With this, we will later be able to
selectively "exclude" buildings from the obstacle map (e.g. to path to
them for entering).

This requires now to pass an "id" field with each building
entry in setpathbuildings (but that is present anyway if the
GSP getbuildings output is passed in).