Include "friendly" range in database.
This extends the database layer for characters and buildings to track
also the longest range (if any) of friendly attacks they have (in addition
to the longest normal attack that is already there).  Furthermore, the
FighterTable's ProcessWithAttacks now includes fighters that have only
friendly attacks.

With this low-level change done, it will be easy to handle proper
processing of friendly attacks in the main, higher-level combat code.