Implement encoding of waypoints.
This implements logic to encode a list of waypoints (which may be rather
large for long paths) into a compressed string, based on xayautil's new
CompressJson functionality.

The findpath RPC method also returns now the encoded string in addition
to the cleartext waypoints, but otherwise the new code is not yet
used anywhere.  It will later replace move parsing (i.e. set-waypoint
moves will need to set an encoded string).