Decode waypoints when parsing moves.
In the move parser, expect the waypoints of characters to be in
encoded form rather than a plain JSON array.  A special case is added
where an explicit JSON "null" value is also valid, and means empty
waypoints.  This is useful as those have a separate meaning, namely
to stop any movement or mining activity.