Add flag for friendly targets to database.
This adds a new flag for characters and buildings in the database
(schema and wrapper classes), which indicates whether or not friendlies
are in range of friendly attacks.  If that is the case, then they
need to be processed for damaging, just as if they had a normal target.

For now, the flag is never set, but it will be set in target finding
later on, so that friendly attacks work properly.