Define "mentecon" effect and fitment.
This defines a new combat effect, namely a flag (codenamed mentecon)
that disrupts a fighter's enemy/friendly detection.  There is also
a new fitment, "vhf mentecon", which has this effect on the unit's
current target (not AoE).

For now, we just apply the effect, but do not take it into account
in the actual combat system.  That will be done in the next step.