Implement hit/miss chance for attacks.
Support a hit/miss chance for attacks (applying damage, not effects).
The chance is based on both a "target" and "weapon" size.  Small weapons
hit large targets with 100% chance; if the weapon size is larger
than the target size, the hit chance is proportional to that.

Both weapons and targets support also a special property, which just means
"hits always" (unless e.g. modified in the future with a tracking
disruptor).