Implement fitment to boost one's hit chance.
This defines a new fitment, "lf hitext", which boosts the hit chance
for one's own attacks.  (The other variants will be added in with a
future general fitment update from the data sheets.)

Also implements general combat logic to apply a modifier to the hit
chance, which for now is just based on that self boost (but in the
future will also include negative combat effects).