Move encodewaypoints to NonStateRpcServer.
The encodewaypoints RPC method was meant only for testing previously;
it turns out to be useful in general, though, e.g. to combine the paths
between multiple explicitly-chosen waypoints in the frontend.

This change exposes the RPC method through the NonStateRpcServer and
also Charon, which makes it fully available to the frontend.