Support uninitialised accounts in database.
Allow accounts in the database to be uninitialised, which means that
they have not yet chosen a faction.  So far, this is not used by
the game logic (i.e. all accounts that are created will be initialised
at the same time), but in the future, such accounts will be able
to send and receive vCHI (but not do anything else in the game).