Database table and wrapper for money supply.
Add a new database table and associated wrapper class for the total
vCHI money supply.  We will use that to track the burnsale of vCHI
for burnt CHI, and follow the different stages correctly.

This will also be useful in the future for general stats, tracking e.g.
how many vCHI have been earned in the game and things like that.