Bump maximum allowed vCHI amount.
Increase the maximum allowed vCHI amount (in a move, but corresponding also
to the total cap) to 100B.  This is what it will be in the full game.

Similar to item counts, we will use QuantityProduct for any computations
of e.g. "price x quantity", so that there is no issue with int64_t overflows
by using a value larger than 32 bits.