Reset game world for 3rd competition.
Set the start height for mainnet to block 2'128'750 as official
start for the 3rd alpha competition.

This will also stay the start height for the full game, although only
Cubit operations (minting / transfer / burn) will be active in the full
game from the start, and the actual gameplay will be enabled only
at a later block height.