Support passing characters to findpath.
Extend the setpathbuildings RPC (renamed to setpathdata), so that also
characters can be passed in and stored in the findpath map data.  This
allows the frontend (and any other client) to also compute paths that
avoid stationary characters on the map, not just buildings.

The character data passed must be a JSON array of objects, containing
at least "faction" and "position".  It can be the full result of
getcharacters.