Define UnblockSpawns fork.
Define a new type of fork and set the heights for it.  The fork will
help to solve the issue of blocked spawn areas in the 0.3 competition
(as well as just test doing a fork in practice).  For now, it has no
effect, but they will be added in later commits.

The fork activates on height 500 on regtest and 2'159'000 on mainnet.