Spawn in starter city after fork.
Update the game rules after the "unblock spawns" fork so that new
characters will spawn in the starter-city building of their faction,
rather than directly on the map.

This helps to reduce the amount of idle characters in the starter
zones that were created with the f2p "faucet" and never touched; they
will now be inside the building, where they are not as annoying.