Make tests more independent of character spawn.
Update the integration tests to be even more independent of where / how
characters are spawned (most already are), so that they will work also
with characters spawned inside buildings.

Also adjust changeCharacterVehicle to take advantage of a character
that already is inside a building (making it a bit more efficient
in that situation).