Move full "passable" check out of DynObstacles.
Instead of exposing IsPassable (with the logic to check for buildings
and vehicles of a certain faction) in DynObstacles directly, expose the
lower-level data about buildings and vehicles and do the logic whether
or not that means "passable" in the callsites.

This is just a refactor for now, but will make it easier to change the
logic for vehicles that can be passed (but slower) in the future.