Count vehicles in DynObstacles.
Change the internal data in DynObstacles, to actually count
how many vehicles there are on any given tile and faction, rather
than just keeping track of true/false.

For now this does nothing to actually change behaviour of the game
(except for making some setpathdata calls valid that were invalid
before), but it will be used to implement the "unblock spawns"
logic in the future.