Spawn on totally free location.
When spawning on the map (not for new characters, but e.g. after
leaving a building), the old behaviour is to choose any
passable tile (i.e. not a building and not a vehicle of the own
faction, but other factions are fine).

This changes it to exclude any vehicles (even other factions).
Those will not be impassable per-se after implementing the "unblock spawn"
game changes, but we still want to avoid them for nicer game behaviour.