After the fork, spawn on totally free location.
When spawning on the map (not for new characters, but e.g. after
leaving a building), the pre-fork behaviour is to choose any
passable tile (i.e. not a building and not a vehicle of the own
faction, but other factions are fine).

This changes it post-fork to exclude any vehicles (even other factions).
Those are not impassable per-se after the fork, but we still want to
avoid them for nicer game behaviour.