Simplifications to movement.py.
Change the movement.py integration test so that the "blocked path" part
is checked with a building rather than a vehicle that comes across, which
will also hold true independent of the "unblock spawns" change.

Also extend the setWaypoints utility method so that it can set a chosen
speed (optionally), and use that to simplify the chosen speed test part.