Avoid id-as-key for building moves.
This changes the format for building updates:  Instead of using a
JSON object keyed by the building IDs as strings, we just use a
single JSON object or an array of objects for the individual operations,
and set the building ID inside each operation:

  {"b": {"id": 42, "sf": 50}}
  {"b": [{"id": 1, "sf": 50}, {"id": 2, "sf": 10}]}

This makes the format and in particular the processing order more
straight-forward and simple.  In theory it also allows now to do multiple
updates to a single building in one move.

This is a first step for a bigger set of changes.  The main change (and
with higher impact) will be doing the same thing also for characters
and group-updates.