Use arrays in godmode sethp/teleport.
Change the format of godmode sethp and teleport to avoid JSON objects
keyed by stringified IDs, and instead just use JSON arrays with
object entries.  This avoids the need to explicitly parse an ID from
a JSON string, and it also removes any doubt about the order in which
the commands might be processed.

The new format for teleport is this:

  {"god": {"teleport": [{"id": 42, "pos": {"x": 1, "y": 2}}, ...]}}

The new format for sethp (buildings and characters) is:

  {"god": {"sethp": {"c": [{"id": 42, "a": 5, ...}, ...], "b": [...]}}}