Array format for character update moves.
This changes the move format for character updates:  Instead of a JSON
object that is keyed by stringified character IDs, we have a JSON array
of individual updates (or just a JSON object for a single update).
Each entry is a JSON object which explicitly specifies the ID as a single
integer or array of integers (for batched updates).

With this form, the order in which the changes are done is more explicit,
and can also be chosen freely by the user.  In fact, it is even possible to
do changes to a single character multiple times, in a particular order and
relative to other character changes.  This will allow better control
for convoy moves, for instance.

The new format is like this:

  {"c": [{"id": 42, ...}, {"id": [10, 20], ...}, ...]}
  {"c": {"id": 1, ...}}