Support waypoint extension moves.
This implements a new type of move, which will *add to* existing
waypoints of a character:

  {"c": {"id": 1, "wpx": "..."}}

The format of "wpx" is an encoded list of waypoints, as for normal
"wp" moves.  The move is valid only if the character has already some
waypoints set, i.e. is already moving.  In this case, all waypoints
from the new "wpx" list will be added at the end of the current
list of waypoints.

This is mostly useful for efficient convoy movements:  For them,
we want a set of characters to move along the same path; but since
they are likely on different initial positions, we first need individual
paths for all of them to some shared first waypoint.  After that, we
want to add the rest of the path for all of them at once.  This can be
done with a move like this:

  {
    "c":
      [
        {"id": 1, "wp": "from position 1 to some shared coordinate"},
        {"id": 2, "wp": "from position 2 to some shared coordinate"},
        {"id": 3, "wp": "from position 3 to some shared coordinate"},
        {"id": [1, 2, 3], "wpx": "long path from that coordinate to target"}
      ]
  }