Implement "game start" fork.
The burnsale that started with the third competition is live and will
not be reset, so that the starting block heights for Taurion are frozen.

This implements a new fork (GameStart), which will be used to specify when
the actual game play (i.e. most of the move logic) will start.  Before that
fork, only Cubit operations (burnsale, transfers) are possible, so this
more or less corresponds to the "real" reset of the game world.