Spawn characters in starter cities.
Spawn new characters inside their starter-city buildings, rather than
directly on the map.  This is simpler and more straight-forward.  They
can then immediately exit the building if they want anyway, which leads
to the same result as before (placing them on the map somewhere around
the starter building).

In particular, this prevents cluttering of the spawn area with characters
that may not be used at all, e.g. from free-to-play claims.